requirements.txt
Flask
opencv-python
pytesseract
Pillow
numpy